simplify connection string configuration
