+++ internal/db/db_internal_test.go
@@ -11,95 +11,6 @@ func (o badOverride) Override() string {
 	return "fail"
 }
 
-func TestParseConnStr(t *testing.T) {
-	cases := []struct {
-		desc      string
-		input     string
-		overrides []Overrideable
-		envs      map[string]string
-		want      *minimalConnConfig
-		wantErr   string
-	}{
-		{
-			desc:  "simple DSN",
-			input: "host=localhost user=test",
-			want: &minimalConnConfig{
-				host: "localhost",
-				user: "test",
-				port: 5432,
-			},
-		},
-		{
-			desc:  "simple URL",
-			input: "postgres://test@localhost",
-			want: &minimalConnConfig{
-				host: "localhost",
-				user: "test",
-				port: 5432,
-			},
-		},
-		{
-			desc:  "full URL",
-			input: "postgres://test:secretPassword!!+!@example:7182/foo?sslmode=require",
-			want: &minimalConnConfig{
-				host:     "example",
-				user:     "test",
-				password: "secretPassword!!+!",
-				port:     7182,
-				db:       "foo",
-				sslmode:  "require",
-			},
-		},
-		{
-			desc:      "DSN w/ override",
-			input:     "host=localhost user=test dbname=foo",
-			overrides: []Overrideable{OverrideDBName("bar")},
-			want: &minimalConnConfig{
-				host: "localhost",
-				user: "test",
-				port: 5432,
-				db:   "bar",
-			},
-		},
-		{
-			desc:      "unknown override",
-			input:     "host=localhost",
-			overrides: []Overrideable{badOverride("fail")},
-			wantErr:   "unknown overrideable: db.badOverride=fail",
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.desc, func(t *testing.T) {
-			if c.envs != nil {
-				for k, v := range c.envs {
-					if err := os.Setenv(k, v); err != nil {
-						t.Errorf("could not set env %s=%s", k, v)
-						return
-					}
-				}
-			}
-
-			mcc, err := parseConnStr(c.input, c.overrides...)
-			if c.wantErr == "" {
-				if err != nil {
-					t.Errorf("unexpected error parsing: %v", err)
-				}
-				if got := mcc.DSN(); got != c.want.DSN() {
-					t.Errorf("incorrect result:\ngot\n%s\nwant\n%s", got, c.want.DSN())
-				}
-			} else {
-				if err == nil {
-					t.Errorf("unexpected lack of error")
-				}
-				if got := err.Error(); got != c.wantErr {
-					t.Errorf("incorrect error:\ngot\n%s\nwant\n%s", got, c.wantErr)
-				}
-			}
-		})
-	}
-}
-
 func TestDetermineTLS(t *testing.T) {
 	cases := []struct {
 		desc    string
@@ -189,35 +100,6 @@ func TestDetermineTLS(t *testing.T) {
 				} else if err.Error() != c.wantErr.Error() {
 					t.Errorf("incorrect error:\ngot\n%v\nwant\n%v", got, c.wantErr)
 				}
-			}
-		})
-	}
-}
-
-func TestDSN(t *testing.T) {
-	cases := []struct {
-		desc  string
-		input minimalConnConfig
-		want  string
-	}{
-		{
-			desc: "",
-			input: minimalConnConfig{
-				host:     "localhost",
-				port:     5432,
-				user:     "user",
-				password: "password",
-				db:       "db",
-				sslmode:  "prefer",
-			},
-			want: "host=localhost user=user password=password dbname=db port=5432 sslmode=prefer application_name=timescaledb-parallel-copy",
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.desc, func(t *testing.T) {
-			got := c.input.DSN()
-			if c.want != got {
-				t.Errorf("wanted %s got %s", c.want, got)
 			}
 		})
 	}
