+++ pkg/csvcopy/csvcopy_test.go
@@ -67,7 +67,7 @@ func TestWriteDataToCSV(t *testing.T) {
 
 	writer.Flush()
 
-	copier, err := NewCopier(connStr, "test-db", "public", "metrics", "CSV", ",", "", "", "device_id,label,value", false, 1, 1, 0, 5000, true, 0, false)
+	copier, err := NewCopier(connStr, "public", "metrics", "CSV", ",", "", "", "device_id,label,value", false, 1, 1, 0, 5000, true, 0, false)
 	require.NoError(t, err)
 
 	reader, err := os.Open(tmpfile.Name())
@@ -154,7 +154,7 @@ func TestErrorAtRow(t *testing.T) {
 
 	writer.Flush()
 
-	copier, err := NewCopier(connStr, "test-db", "public", "metrics", "CSV", ",", "", "", "device_id,label,value", false, 1, 1, 0, 2, true, 0, false)
+	copier, err := NewCopier(connStr, "public", "metrics", "CSV", ",", "", "", "device_id,label,value", false, 1, 1, 0, 2, true, 0, false)
 	require.NoError(t, err)
 	reader, err := os.Open(tmpfile.Name())
 	require.NoError(t, err)
